Fix service rescheduling on clock skew/timeperiod change
This patch ensures that services and hosts are never scheduled one
year into the future and set to never be rescheduled again.

Previously, this could happen if the next preferred time happened
to already be valid, but stops being so because of clock skew or
someone changing the timeperiod definition between two Nagios
restarts while retaining scheduling info.

Patch-sent-by: Ricardo Maraschini <ricardo.maraschini@opservices.com.br>
Signed-off-by: Andreas Ericsson <ae@op5.se>