nebmods: Protect against callbacks removing themselves
Previously a callback deregistering itself from the chain of
callbacks to run at the event-type the callback registered
as (err...), Nagios would segfault since the pointer to the
next callback couldn't be looked up.

This patch takes care of it, making it possible for a module
to unload the callback it's currently running.

Signed-off-by: Andreas Ericsson <ae@op5.se>


Author: Andreas Ericsson <ae@op5.se>