Add a make-target to indent all sources
Using astyle + find, this implements the style discussed in the
thread at http://comments.gmane.org/gmane.network.nagios.devel/7683
which is sort of what we agreed on at the first (very informal)
Nagios core dev conference in Sweden in 2008.

"make indent" or "make pretty" should enforce the rules to all
sources located in the checked out tree. It's probably wise to
run it after applying a patch and recommend patch submitters to
run it before creating their patches. The fact that it's a
readily available make target makes it quite simple to add janitor
commits, re-fixing any broken indentation every once in a while
though.

Signed-off-by: Andreas Ericsson <ae@op5.se>