core: Try to avoid a single check being reaped multiple times
Previously, a check that happened to time out at the same time
as it was completing could lead to the checkresult output file
being either written to after it's been closed, or closed more
than once.

This patch should close the window where it's possible for that
to happen, thus preventing coredumps from occurring with the
distressing frequency at least one of our customers has seen.

Signed-off-by: Andreas Ericsson <ae@op5.se>