base/notifications: Plug memory leak and (micro-)optimize the code
Since we moved to using thread-safe macro management in the
notification API, we've been suffering from a small memory leak of
about 300-400 bytes per notification-generating alert (not per
notification). The earlier code left the notification macros in
place to be cleared out the next time we needed to actually send a
notification, but the stackbased macro structure has no such luxury
and the necessary free() calls weren't in place after the data was
allocated.

Instead of copying the strings into the macro structure and keeping
track of them there, we just assign them by reference and only free()
the numerical macros, which are generated by asprintf() and are thus
heap-located. This provides a small improvement in memory and cpu usage
when sending notifications and will also cause less memory
fragmentation.

Signed-off-by: Andreas Ericsson <ae@op5.se>