core: Be more aggressive in deciding check rescheduling on startup
Previously we ignored rescheduling checks if they already had a
next_check time in the future read from retention data. This patch
changes it so that we ignore rescheduling checks if they have a
next_check time not too far in the future. Not too far is decided
by checking that 'next_check < now+(check_interval*interval_length)'

This is necessary to make checks that have had their timeperiod
changed during the restart period not be scheduled several weeks
in the future, which might happen when people fsck up some of
their more advanced timeperiods.

Signed-off-by: Andreas Ericsson <ae@op5.se>