core: Read last_program_stop from retention data
We know that the retention data file gets written on shutdown, so
we use the creation_time in the retention file when determining it.
This isn't necessarily very accurate, but it's good enough for the
purposes we'll want to put it to. At least assuming Nagios didn't
crash the last time it was running, but if we start assuming bugs
everywhere we'll never get anything done.

Signed-off-by: Andreas Ericsson <ae@op5.se>