core: Use last_program_stop when deciding if checks are fresh or not
If a check was last done passively, we assume it's going to continue
that way and we need to handle the fact that Nagios might have been
shut off for quite a long time. If so, we mustn't spam freshness
notifications but use event_start instead of last_check to determine
freshness expiration time.

The threshold for "long time" is determined as 61.8% of the normal
freshness threshold based on vast heuristical research (ie, "some
guy once told me the golden ratio is good for loads of stuff").

Signed-off-by: Andreas Ericsson <ae@op5.se>