Avoid coredumps on list traversion
When deleting downtime or comments based on certain filtering
options (such as hostname, service description, start time or
whatnot), we used to traverse the lists using pointers to
already deleted objects when we found anything to delete. For
most cases, the compiler optimized this away for us, but with
some settings it's quite possible that this will result in us
attempting to read memory that belongs to a different program,
resulting in a SIGSEGV and a subsequent crash and coredump.

The "make indent" run after applying the two patches resulted
in a small unrelated indentation fix.

Fixed-by: Dorian Guillois <dguillois@merethis.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>